requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.35.0
faiss-cpu>=1.7.4

# Image processing
pillow>=10.0.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Async HTTP
aiohttp>=3.9.0
aiofiles>=23.2.0

# Data
numpy>=1.24.0
pydantic>=2.0.0

# Progress & CLI
tqdm>=4.66.0
click>=8.1.0

# AVIF support
pillow-avif-plugin>=1.5.0
opencv-python-headless>=4.8.0
scipy>=1.11.0
